Created a quiz using react.js and next.js
In this quiz their are 10 questions for each correct answer 1 point is rewarded after completion of quiz the score is shown and a re-attempt option is show also the user cannot attend the quiz of it is not in full screen. The time limit of the quiz is 10 min and the time will not restart even after the reloading the webpage and closing the tab.